Remove unnecessary forward declarations and put static_factory functionality in core module
