+++ libs/core/runtime_configuration/include/hpx/runtime_configuration/static_factory_data.hpp
@@ -25,12 +25,27 @@ namespace hpx { namespace components {
         hpx::util::plugin::get_plugins_list_type get_factory;
     };
 
-    HPX_EXPORT void init_registry_module(static_factory_load_data_type const&);
-    HPX_EXPORT void init_registry_factory(static_factory_load_data_type const&);
-    HPX_EXPORT void init_registry_commandline(
-        static_factory_load_data_type const&);
-    HPX_EXPORT void init_registry_startup_shutdown(
-        static_factory_load_data_type const&);
+    HPX_CORE_EXPORT bool& get_initial_static_loading();
+
+    HPX_CORE_EXPORT std::vector<static_factory_load_data_type>&
+    get_static_module_data();
+    HPX_CORE_EXPORT
+    void init_registry_module(static_factory_load_data_type const&);
+
+    HPX_CORE_EXPORT bool get_static_factory(
+        std::string const& instance, util::plugin::get_plugins_list_type& f);
+    HPX_CORE_EXPORT
+    void init_registry_factory(static_factory_load_data_type const&);
+
+    HPX_CORE_EXPORT bool get_static_commandline(
+        std::string const& instance, util::plugin::get_plugins_list_type& f);
+    HPX_CORE_EXPORT
+    void init_registry_commandline(static_factory_load_data_type const&);
+
+    HPX_CORE_EXPORT bool get_static_startup_shutdown(
+        std::string const& instance, util::plugin::get_plugins_list_type& f);
+    HPX_CORE_EXPORT
+    void init_registry_startup_shutdown(static_factory_load_data_type const&);
 }}    // namespace hpx::components
 
 ///////////////////////////////////////////////////////////////////////////////
