+++ libs/full/init_runtime/include/hpx/hpx_init_impl.hpp
@@ -14,6 +14,7 @@
 #include <hpx/hpx_main_winsocket.hpp>
 #include <hpx/hpx_user_main_config.hpp>
 #include <hpx/init_runtime/detail/run_or_start.hpp>
+#include <hpx/init_runtime_local/init_runtime_local.hpp>
 #include <hpx/modules/program_options.hpp>
 #include <hpx/prefix/find_prefix.hpp>
 #include <hpx/runtime_configuration/runtime_mode.hpp>
