+++ libs/full/init_runtime/include/hpx/init_runtime/detail/run_or_start.hpp
@@ -17,9 +17,6 @@ namespace hpx {
             util::function_nonser<int(
                 hpx::program_options::variables_map& vm)> const& f,
             int argc, char** argv, init_params const& params, bool blocking);
-
-        HPX_CORE_EXPORT int init_helper(hpx::program_options::variables_map&,
-            util::function_nonser<int(int, char**)> const&);
     }    // namespace detail
     /// \endcond
 }    // namespace hpx
